voice-call(msteams): share MsteamsVideoFrame type, validate TTS merge, surface audio.frame send failures
